Shoot guns (#2)
* Started moving stuff into separate systems

* Moved startup systems

* Lots of refactoring and structuring... but also...
added a "trigger_pulled" property to the PlayerControl component.

* More refactoring.

Adds the trigger_pulled component that will be added when
space is pressed - ah, but what if we already have that component?
Should we do this in an entirely different system? Whaaat?

I think the code crashes as it is written now. Fudge.

* More refactorings, of course
Adds a weapon_def and a weapon instance.
Should these be cloned? Should they reference the same instance?
What is the Rc<T> in Rust?

Adds a shooting system.

* WIP on shooting, creating projectiles at the moment

* Almost there, yay, yay...

* Using arc is probably fucking insane, mate.

* Can change weapons and stuff, but no shooting occurs. Perhaps we
need a little sprite?

* Remove the boids so we have the time to shoot some.

* We have SLOW bullets!

* The bullets are fast now!

* Fixed stuff relating to events for collisions

* Added events package
Added events for major events we have right now

* Adds effect component and system for tracking health of boids

* We can now with confidence say that we can shoot stuff.